Use the gutentag gem to handle tagging for me
I tried to write tagging myself, and while it was an experience that
teached me a lot, the thing it taught me in the first place was to
not try to write a tagging library in the first place.

There is yet some cleanup to be done, which will be done in a separate
commit to be able to tear them apart in the future, should it be
needed.